added rudimentary userIds so that it won't give you a question you wrote

+++ grails-app/controllers/bucketquestions/BucketController.groovy
@@ -6,6 +6,7 @@ class BucketController {
     Random rand = new Random()
 
     def index(String codeword) {
+        int userId
         if(codeword == null || codeword.isEmpty()) {
             //if no codeword is given redirect to main index
             redirect(uri:"/")
@@ -13,11 +14,16 @@ class BucketController {
         Bucket b = Bucket.findByCodeword(codeword)
         if (b == null) {
             //if a codeword is given that doesn't match an existing bucket, make a new one
-            b = new Bucket(codeword: codeword)
+            b = new Bucket(codeword: codeword, userId: 1)
+            b.save()
+            userId = 0
+        } else {
+            userId = b.userId   //todo: save userId in a cookie and check that cookie before giving them a new one
+            b.setUserId(b.userId+1)
             b.save()
         }
 
-        render(view:"index", model:[bucket:b, codeword:codeword])
+        render(view:"index", model:[bucket:b, codeword:codeword, userId: userId])
     }
 
     def addQuestion(){
@@ -25,11 +31,12 @@ class BucketController {
         boolean success = false
         String codeword = request.JSON["codeword"]
         String questionText = request.JSON["questionText"]
+        String userId = request.JSON["userId"]
         if(codeword != null && !codeword.isEmpty()
                 && Bucket.countByCodeword(codeword) > 0){
             Bucket b = Bucket.findByCodeword(codeword)
             if(questionText != null && !questionText.isEmpty()) {
-                Question q = new Question(questionText: questionText)
+                Question q = new Question(questionText: questionText, createdBy: userId)
                 b.addToQuestions(q)
                 b.save()
             }
@@ -44,6 +51,7 @@ class BucketController {
 
     def getQuestion(){
         String codeword = request.JSON["codeword"]
+        String userId = request.JSON["userId"]
         String question
         int numQuestions = 0
         if(codeword == null || codeword.isEmpty()){
@@ -52,10 +60,15 @@ class BucketController {
             Bucket b = Bucket.findByCodeword(codeword)
             if (b != null) {
                 if (b.questions.size() > 0) {
-                    Question q = b.questions[rand.nextInt(b.questions.size())] as Question
-                    question = "Current Question: " + q.questionText
-                    b.removeFromQuestions(q)
-                    q.delete()
+                    def validQuestions = b.questions.findAll {it.createdBy != userId}
+                    if(validQuestions.size() > 0) {
+                        Question q = validQuestions[rand.nextInt(b.questions.size())] as Question
+                        question = "Current Question: " + q.questionText
+                        b.removeFromQuestions(q)
+                        q.delete()
+                    } else {
+                        question = "Sorry, but you can't answer a question you wrote and you wrote all the remaining questions!"
+                    }
                     numQuestions = b.questions.size()
                 } else {
                     question = "There aren't any questions in this bucket. Use the form below to create more!"
